Moved zobrist keys to specific file
