+++ perft
@@ -16,8 +16,26 @@ arg="-g"
 if [[ "$1" == "time" || "$1" == "test" ]]; then
     arg="-O3"
 fi
+
+VAR="testing/perft.c 
+testing/logChessStructs.c 
+
+src/chessGameEmulator.c 
+src/moveGenerator.c 
+
+src/utils/fenString.c 
+src/utils/utils.c 
+
+src/state/board.c 
+src/state/gameState.c 
+src/state/zobristKey.c 
+
+src/magicBitBoard/magicBitBoard.c 
+src/magicBitBoard/rook.c 
+src/magicBitBoard/bishop.c"
+
 # For checking memory bugs, add option -fsanitize=address 
-gcc -Wall -Wextra -Werror -Wunused "$arg" -o perftTesting testing/perft.c testing/logChessStructs.c src/chessGameEmulator.c src/moveGenerator.c src/utils/fenString.c src/utils/utils.c src/state/board.c src/state/gameState.c src/magicBitBoard/magicBitBoard.c src/magicBitBoard/rook.c src/magicBitBoard/bishop.c 
+echo -n "$VAR" | tr -d '\n' | xargs gcc -Wall -Wextra -Werror -Wunused "$arg" -o perftTesting
 
 if [ $? -ne 0 ]; then
     exit 1
